parallelization: treat off-diagonal matrix entries on the process boundaries properly
this is required so that the element-centered finite volume method
does not handle each partition of the domain separately. (i.e. so that
fluxes accross faces on the process boundaries are considered) mea culpa!

The fix for this is to also include these entries in the matrix which
uses domestic indices. This required some rather extensive changes to
the blacklisting mechanism as for this it must be possible to
translate the index of a blacklisted entity (i.e., an entity in a
ghost or an overlap cell) to a domestic index (i.e., the corresponding
index in the algebraic overlap).

Note that the code for algebraic overlaps is *fun* and the person who
wrote it should be tarred and feathered. (*ouch* ;)) Seriously: Better
approaches than "lets-throw-this-away-and-use-grid-overlaps" are
deeply appreciated. (The grid overlap is not really useful in Dune
because only "Mickey Mouse grids" like Dune::YaspGrid support it.)

+++ ewoms/linear/globalindices.hh
@@ -184,7 +184,7 @@ public:
         std::cout << "(domestic index, global index, domestic->global->domestic)"
                   << " list for rank " << myRank_ << "\n";
 
-        for (int domIdx = 0; domIdx < domesticToGlobal_.size(); ++domIdx)
+        for (size_t domIdx = 0; domIdx < domesticToGlobal_.size(); ++domIdx)
             std::cout << "(" << domIdx << ", " << domesticToGlobal(domIdx)
                       << ", " << globalToDomestic(domesticToGlobal(domIdx)) << ") ";
         std::cout << "\n" << std::flush;
